Fix FFM parameter types for texture creation
The Swift function metallum_create_texture_2d expects all numeric
parameters as 64-bit (LONG), but Java code was using 32-bit (INT).

Fixed FunctionDescriptor for:
- createTexture2D: device, pixelFormat(LONG), width, height, depthOrLayers,
  mipLevels, cubeCompatible, usage(LONG), storageMode(LONG), label
- createTexture3D / createTextureCube: same signature

This was causing the MTLResourceOptions validation error because
the wrong bits were being passed to Metal.

+++ common/src/main/java/net/irisshaders/iris/metal/bridge/IrisMetalNativeBridge.java
@@ -281,14 +281,16 @@ public final class IrisMetalNativeBridge {
         // 纹理
         // metallum_create_texture_2d 需要 device 参数
         // 参数顺序: device, pixelFormat, width, height, depthOrLayers, mipLevels, cubeCompatible, usage, storageMode, label
+        // 注意: pixelFormat, usage, storageMode 都是 LONG 类型！
         createTexture2D = downcall(lookup, "metallum_create_texture_2d",
-                FunctionDescriptor.of(ValueLayout.ADDRESS, ValueLayout.ADDRESS, INT, LONG, LONG, LONG, LONG, LONG, INT, INT, ValueLayout.ADDRESS));
+                FunctionDescriptor.of(ValueLayout.ADDRESS, ValueLayout.ADDRESS, LONG, LONG, LONG, LONG, LONG, LONG, LONG, LONG, ValueLayout.ADDRESS));
         // Note: metallum_create_texture 需要 device 参数，用于创建 3D 和 cube 纹理
         // 参数顺序: device, pixelFormat, width, height, depthOrLayers, mipLevels, dimension, cubeCompatible, usage, storageMode, label
+        // 注意: pixelFormat, usage, storageMode 都是 LONG 类型！
         createTexture3D = optionalDowncall(lookup, "metallum_create_texture",
-                FunctionDescriptor.of(ValueLayout.ADDRESS, ValueLayout.ADDRESS, INT, LONG, LONG, LONG, LONG, LONG, LONG, INT, INT, ValueLayout.ADDRESS));
+                FunctionDescriptor.of(ValueLayout.ADDRESS, ValueLayout.ADDRESS, LONG, LONG, LONG, LONG, LONG, LONG, LONG, LONG, LONG, ValueLayout.ADDRESS));
         createTextureCube = optionalDowncall(lookup, "metallum_create_texture",
-                FunctionDescriptor.of(ValueLayout.ADDRESS, ValueLayout.ADDRESS, INT, LONG, LONG, LONG, LONG, LONG, LONG, INT, INT, ValueLayout.ADDRESS));
+                FunctionDescriptor.of(ValueLayout.ADDRESS, ValueLayout.ADDRESS, LONG, LONG, LONG, LONG, LONG, LONG, LONG, LONG, LONG, ValueLayout.ADDRESS));
         textureReplaceRegion = optionalDowncall(lookup, "metallum_texture_replace_region",
                 FunctionDescriptor.ofVoid(ValueLayout.ADDRESS, ValueLayout.ADDRESS, LONG, LONG, LONG, LONG, LONG, LONG, LONG, LONG, INT, LONG));
         textureGetBytes = optionalDowncall(lookup, "metallum_texture_get_bytes",
@@ -670,9 +672,10 @@ public final class IrisMetalNativeBridge {
 
     // ===== 纹理 =====
     // metallum_create_texture_2d 需要 device 参数
-    public static MemorySegment createTexture2D(int pixelFormat, long width, long height,
+    // 注意: pixelFormat, usage, storageMode 在 Swift 中是 64 位的
+    public static MemorySegment createTexture2D(long pixelFormat, long width, long height,
                                                  long depthOrLayers, long mipLevels, long cubeCompatible,
-                                                 int usage, int storageMode, String label) {
+                                                 long usage, long storageMode, String label) {
         try (Arena arena = Arena.ofConfined()) {
             MemorySegment device = getSystemDefaultDevice();
             MemorySegment labelSeg = allocateUtf8String(label);
@@ -688,14 +691,14 @@ public final class IrisMetalNativeBridge {
         }
     }
 
-    public static MemorySegment createTexture3D(int pixelFormat, long width, long height,
-                                                 long depth, long mipLevels, int usage, int storageMode, String label) {
+    public static MemorySegment createTexture3D(long pixelFormat, long width, long height,
+                                                 long depth, long mipLevels, long usage, long storageMode, String label) {
         if (createTexture3D == null) {
             return MemorySegment.NULL;  // 返回 NULL 而不是抛异常，让调用者处理
         }
         try (Arena arena = Arena.ofConfined()) {
             MemorySegment device = getSystemDefaultDevice();
-            MemorySegment labelSeg = allocateUtf8String( label);
+            MemorySegment labelSeg = allocateUtf8String(label);
             // dimension=3 表示 3D 纹理
             return (MemorySegment) createTexture3D.invoke(device, pixelFormat, width, height, depth, mipLevels, 3, 0, usage, storageMode, labelSeg);
         } catch (Throwable t) {
@@ -703,14 +706,14 @@ public final class IrisMetalNativeBridge {
         }
     }
 
-    public static MemorySegment createTextureCube(int pixelFormat, long size, long mipLevels,
-                                                   int usage, int storageMode, String label) {
+    public static MemorySegment createTextureCube(long pixelFormat, long size, long mipLevels,
+                                                   long usage, long storageMode, String label) {
         if (createTextureCube == null) {
             return MemorySegment.NULL;  // 返回 NULL 而不是抛异常，让调用者处理
         }
         try (Arena arena = Arena.ofConfined()) {
             MemorySegment device = getSystemDefaultDevice();
-            MemorySegment labelSeg = allocateUtf8String( label);
+            MemorySegment labelSeg = allocateUtf8String(label);
             // cubeCompatible=1 表示 cube 纹理, dimension=2 表示 2D
             return (MemorySegment) createTextureCube.invoke(device, pixelFormat, size, 1, mipLevels, 2, 1, usage, storageMode, labelSeg);
         } catch (Throwable t) {
